adapt3r/envs/mimicgen/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mimicgen"
version = "0.1.0"
description = "MimicGen: A framework for generating and manipulating motion data"
authors = [
    { name = "Ajay Mandlekar", email = "amandlekar@nvidia.com" },
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "h5py>=3.8.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "hydra-core>=1.3.0",
    "gymnasium>=0.28.0",
    "robosuite>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.isort]
known-first-party = ["mimicgen"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["torch.*", "numpy.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=mimicgen"

[tool.hatch.build.targets.wheel]
packages = ["mimicgen"]
